LU-5 readdir read multiple pages per rpc
add support for readdir to read multiple pages per rpc:
* because client has no idea how many directory pages it can read, it
  tries to read maximum pages each time, but will only store pages
  read from mds into page cache.
* add a flag LDF_COLLIDE to mark a dir page hash collides with the
  next page, and client will remove this page from page cache after
  processing.
* upon readpage bulk io failure, client won't be evicted, and client
  will resend the bulk request.
* support large page size on client: MDS_READPAGE RPC will fill reply
  with page size LU_PAGE_SIZE(4k), and if client page is bigger than
  LU_PAGE_SIZE, several pages will be integrated into one dir page with
  CFS_PAGE_SIZE.

+++ lustre/tests/sanity.sh
@@ -793,7 +793,17 @@ test_24v() {
 
 	mkdir -p $DIR/d24v
 	createmany -m $DIR/d24v/$tfile $NRFILES
+
 	ls $DIR/d24v >/dev/null || error "error in listing large dir"
+
+	cancel_lru_locks mdc
+	lctl set_param mdc.*.stats clear
+	/bin/ls $DIR/d24v >/dev/null || error "error in listing large dir"
+	mds_readpage=`lctl get_param mdc.*.stats | awk '/^mds_readpage/ {print $2}'`
+	# LU-5
+	# it's 672 dir pages total, but with large readdir, 3 RPC can fetch all,
+	# and two extra RPC for mountpoint and $DIR
+	[ $mds_readpage -gt 5 ] && error "large readdir doesn't take effect"
 
 	rm $DIR/d24v -rf
 }
